I apply the longitude-nonlinearity to my plots
The OSM tiles have a nonlinear relationship between longitude and tile pixels.
Previously I would linearize it, but now the gnuplot script generated here
applies the nonlinearity to the y2 axis, so the plotted image is not distorted,
but the axes are still showing the correct data. A side-effect of this is that
latitude is on the y2 axis and longitude is on the x2 axis.